requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dataclasses-json==0.5.8
dnspython==2.3.0
ffmpeg-python==0.2.0
filelock==3.12.2
Flask==2.3.2
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
huggingface==0.0.1
huggingface-hub==0.15.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.206
langchainplus-sdk==0.0.16
llvmlite==0.40.1rc1
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
ndg-httpsclient==0.5.1
networkx==3.1
nltk==3.8.1
numba==0.57.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openai-whisper==20230314
openapi-schema-pydantic==1.2.4
pandas==2.0.2
Pillow==9.5.0
pinecone-client==2.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pymongo==4.4.0
PyMuPDF==1.22.5
pyOpenSSL==23.2.0
pypdf==3.10.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
SQLAlchemy==2.0.16
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.1
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2